+++ misc/get_estimates_unsuppressed_proportion_non_participants_vl200.R
@@ -2,12 +2,26 @@ library(data.table)
 library(ggplot2)
 library(Hmisc)
 library(rstan)
+library(here)
 
 # directory of the repository
-indir.repository <- getwd()
+gitdir <- here()
+source(file.path(gitdir, "paths.R"))
+
+# TODO: shozen: do you think this would be helpful? 
+# library(optparse)
+# option_list <- list(
+#     make_option(
+#         "--outdir",
+#         type = "",
+#         default = ,
+#         help = "",
+#         dest= ""
+#     ),
+# )
+# args <- parse_args(OptionParser(option_list = option_list))
 
 # outdir directory for stan fit
-indir.deepsequence.analyses <- '~/Box\ Sync/2021/ratmann_deepseq_analyses/live'
 if (dir.exists(indir.deepsequence.analyses)) {
   outdir <- file.path(indir.deepsequence.analyses, 'PANGEA2_RCCS', 'vl_suppofinfected_by_gender_loc_age')
 } else {
@@ -17,11 +31,11 @@ if (dir.exists(indir.deepsequence.analyses)) {
 
 
 # files
-path.stan <- file.path(indir.repository, 'misc', 'stan_models', 'binomial_gp.stan')
-path.data <- file.path(indir.repository, 'data', 'aggregated_newlyregistered_count_unsuppressed_vl200.csv')
+path.stan <- file.path(gitdir, 'misc', 'stan_models', 'binomial_gp.stan')
+path.data <- file.path(gitdir, 'data', 'aggregated_newlyregistered_count_unsuppressed_vl200.csv')
 
 # Load count of newly registered participants with unsuppressed viral loads
-vla <- as.data.table( read.csv(path.data) )
+vla <- fread(path.count.newly.unsupp.vl200)
 
 
 ##########
@@ -274,8 +288,15 @@ stats[['max_rhat']] = convergence[, round(max(rhat), 4)]
 
 #########
 
-file.name <- file.path(indir.repository, 'fit', paste0('RCCS_nonsuppressed_proportion_posterior_samples_vl_200_newlyregistered_221121.rds'))
-saveRDS(nsinf.samples, file = file.name)
+# file.name <- file.path(gitdir.fit, 'RCCS_nonsuppressed_proportion_posterior_samples_vl_200_newlyregistered_221121.rds')
+file.name <- file.unsuppressedviralload.newly.vl200 
+if(! file.exists(file.name))
+{
+    cat("Saving file:", file.name, '\n')
+    saveRDS(nsinf.samples, file = file.name)
+}else{
+    cat("File:", file.name, "already exists...\n")
+}
 
-file.name <- file.path(outdir, paste0('RCCS_nonsuppressed_proportion_model_fit_newlyregistered_vl200_221121.RDS'))
+file.name <- file.path(outdir, 'RCCS_nonsuppressed_proportion_model_fit_newlyregistered_vl200_221121.RDS')
 saveRDS(stats, file = file.name)
