+++ misc/get_prevalence_share_sex.R
@@ -2,16 +2,30 @@ library(data.table)
 library(ggplot2)
 require(lubridate)
 library(dplyr)
+library(here)
 
-# directory to reposity
-indir.repository <- getwd()
+# directory of the repository
+gitdir <- here()
+source(file.path(gitdir, "paths.R"))
 
-# files
-file.prevalence <- file.path(indir.repository, 'fit', paste0('RCCS_prevalence_posterior_sample_221116.rds'))
-file.eligible.count <- file.path(indir.repository, 'data', 'RCCS_census_eligible_individuals_221116.csv')
+# TODO: shozen: do you think this would be helpful? 
+# library(optparse)
+# option_list <- list(
+#     make_option(
+#         "--outdir",
+#         type = "",
+#         default = ,
+#         help = "",
+#         dest= ""
+#     ),
+# )
+# args <- parse_args(OptionParser(option_list = option_list))
+
+file.exists(file.eligible.count) |> stopifnot()
+file.exists(file.prevalence) |> stopifnot()
 
 # load census eligible ount
-eligible_count <- as.data.table(read.csv(file.eligible.count))
+eligible_count <- fread(file.eligible.count)
 
 # load proportion prevalence
 proportion_prevalence <- as.data.table(readRDS(file.prevalence))
@@ -93,7 +107,12 @@ ggplot(sing, aes(x = ROUND)) +
 #########################################
 
 tmp <- merge(sing.age, sing, by=c('ROUND', 'COMM', 'SEX'))
-
-file.name <- file.path(indir.repository, 'fit', paste0('RCCS_prevalence_share_sex_221116.csv'))
-write.csv(tmp, file = file.name, row.names = FALSE)
+file.name <- file.prevalence.share
+if( ! file.exists(file.name))
+{
+    cat('Saving file:', file.name, '...\n')
+    write.csv(tmp, file = file.name, row.names = FALSE)
+}else{
+    cat('File', file.name, 'already exists.\n')
+}
