+++ misc/get_treatment_cascade_non_participants_vl200.R
@@ -2,12 +2,26 @@ library(data.table)
 library(ggplot2)
 require(lubridate)
 library(dplyr)
+library(here)
 
-# directory to repository
-indir.repository <- getwd()
+# directory of the repository
+gitdir <- here()
+source(file.path(gitdir, "paths.R"))
+
+# TODO: shozen: do you think this would be helpful? 
+# library(optparse)
+# option_list <- list(
+#     make_option(
+#         "--outdir",
+#         type = "",
+#         default = ,
+#         help = "",
+#         dest= ""
+#     ),
+# )
+# args <- parse_args(OptionParser(option_list = option_list))
 
 # outdir to save figures
-indir.deepsequence_analyses <- '~/Box\ Sync/2021/ratmann_deepseq_analyses/live/'
 if (dir.exists(indir.deepsequence_analyses)) {
   outdir <- file.path(indir.deepsequence_analyses, 'PANGEA2_RCCS', 'treatment_cascade_by_gender_loc_age')
 } else {
@@ -15,12 +29,9 @@ if (dir.exists(indir.deepsequence_analyses)) {
   if (!dir.exists(outdir)) dir.create(outdir);
 }
 
-# posterior samples
-file.unsuppressedviralload <- file.path(indir.repository, 'fit', paste0('RCCS_nonsuppressed_proportion_posterior_samples_vl_200_newlyregistered_221121.rds'))
-file.selfreportedart <- file.path(indir.repository, 'fit', paste0('RCCS_art_posterior_samples_newlyregistered_vl200_221208.rds'))
-
-# specificity and sensitivity art reporting
-file.spec.sens.art = file.path(indir.repository, 'data', 'sensitivity_specificity_art_vl200.csv')
+file.exists(file.unsuppressedviralload.newly.vl200) |> stopifnot()
+file.exists(file.selfreportedart.newly.vl200 )|> stopifnot()
+file.exists(file.spec.sens.art.vl200 )|> stopifnot()
 
 ps <- c(0.025,0.5,0.975)
 qlab <- c('CL','M','CU')
@@ -178,8 +189,20 @@ stopifnot(nrow(ns[COMM == 'fishing']) == ns[, length(unique(AGEYRS))] * ns[, len
 
 ####################################
 
-file.name <- file.path(indir.repository, 'fit', paste0('RCCS_treatment_cascade_nonparticipants_posterior_samples_vl200_221208.rds')) 
-saveRDS(df, file = file.name)
+file.name <- file.treatment.cascade.prop.nonparticipants.vl200.samples 
+if(! file.exists(file.name))
+{
+    cat("Saving file:", file.name, '\n')
+    saveRDS(df, file = file.name)
+}else{
+    cat("File:", file.name, "already exists...\n")
+}
 
-file.name <- file.path(indir.repository, 'fit', paste0('RCCS_treatment_cascade_nonparticipants_estimates_vl200_221208.csv'))
-write.csv(ns, file = file.name, row.names = F)
+file.name <- file.treatment.cascade.prop.nonparticipants.vl200 
+if(! file.exists(file.name))
+{
+    cat("Saving file:", file.name, '\n')
+    write.csv(ns, file = file.name, row.names = F)
+}else{
+    cat("File:", file.name, "already exists...\n")
+}
