+++ misc/get_treatment_cascade_participants.R
@@ -2,12 +2,26 @@ library(data.table)
 library(ggplot2)
 require(lubridate)
 library(dplyr)
+library(here)
 
-# directory to repository
-indir.repository <- getwd()
+# directory of the repository
+gitdir <- here()
+source(file.path(gitdir, "paths.R"))
+
+# TODO: shozen: do you think this would be helpful? 
+# library(optparse)
+# option_list <- list(
+#     make_option(
+#         "--outdir",
+#         type = "",
+#         default = ,
+#         help = "",
+#         dest= ""
+#     ),
+# )
+# args <- parse_args(OptionParser(option_list = option_list))
 
 # outdir to save figures
-indir.deepsequence_analyses <- '~/Box\ Sync/2021/ratmann_deepseq_analyses/live/'
 if (dir.exists(indir.deepsequence_analyses)) {
   outdir <- file.path(indir.deepsequence_analyses, 'PANGEA2_RCCS', 'treatment_cascade_by_gender_loc_age')
 } else {
@@ -17,11 +31,9 @@ if (dir.exists(indir.deepsequence_analyses)) {
 
 
 # posterior samples
-file.unsuppressedviralload <- file.path(indir.repository, 'fit',  paste0('RCCS_nonsuppressed_proportion_posterior_samples_vl_1000_220818.rds'))
-file.selfreportedart <- file.path(indir.repository, 'fit', paste0('RCCS_art_posterior_samples_221208.rds'))
-
-# specificity and sensitivity art reporting
-file.spec.sens.art = file.path(indir.repository, 'data', 'sensitivity_specificity_art.csv')
+file.exists(file.unsuppressedviralload) |> stopifnot()
+file.exists(file.selfreportedart) |> stopifnot()
+file.exists(file.spec.sens.art) |> stopifnot()
 
 ps <- c(0.025,0.5,0.975)
 qlab <- c('CL','M','CU')
@@ -182,8 +194,22 @@ stopifnot(nrow(ns[COMM == 'fishing']) == ns[, length(unique(AGEYRS))] * ns[, len
 
 ####################################
 
-file.name <- file.path(indir.repository, 'fit', paste0('RCCS_treatment_cascade_participants_posterior_samples_221208.rds')) #221208b without adjustement for sens/susc
+# file.name <- file.path(gitdir.fit,'RCCS_treatment_cascade_participants_posterior_samples_221208.rds') #221208b without adjustement for sens/susc
+file.name <- file.treatment.cascade.prop.participants.samples 
+if(! file.exists(file.name))
+{
+    cat("Saving file:", file.name, '\n')
+}else{
+    cat("File:", file.name, "already exists...\n")
+}
 saveRDS(df, file = file.name)
 
-file.name <- file.path(indir.repository, 'fit', paste0('RCCS_treatment_cascade_participants_estimates_221208.csv')) #221208b without adjustement for sens/susc
-write.csv(ns, file = file.name, row.names = F)
+# file.name <- file.path(gitdir.fit, 'RCCS_treatment_cascade_participants_estimates_221208.csv') #221208b without adjustement for sens/susc
+file.name <- file.treatment.cascade.prop.participants 
+if(! file.exists(file.name))
+{
+    cat("Saving file:", file.name, '\n')
+    write.csv(ns, file = file.name, row.names = F)
+}else{
+    cat("File:", file.name, "already exists...\n")
+}
